Database and permission patches

+++ tg_bot/modules/sql/antispam_sql.py
@@ -1,13 +1,14 @@
 import threading
 
-from sqlalchemy import Column, UnicodeText, Integer, String, Boolean
+from sqlalchemy import Column, UnicodeText, String, Boolean
+from sqlalchemy.sql.sqltypes import BigInteger
 
 from tg_bot.modules.sql import BASE, SESSION
 
 
 class GloballyBannedUsers(BASE):
     __tablename__ = "gbans"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     name = Column(UnicodeText, nullable=False)
     reason = Column(UnicodeText)
